Add ant's number security(int max), same for room number, cleanng some code
**lem-in.h**
	- Has been updated to support function's modification described
	bellow.

**edmondcarp.c**
	- unused variable removed.

**optimize_ants.c**
	- Comment unused function, probably to be remove later
	- add a to be remove tag

**send_ants.c**
	- remove unused variables from functions.

**solve.c**
	- call to send_ants() has been updated accordingly to previous
	changes.

**get_input.c**
	- add long tmp variable to be able to store a number bigger than
	INT_MAX, if it is, we stop the programm to avoid complications.

**get_room.c**
	- id becomes long (see changes in new_room.c , now, if it reaches more than INT MAX value,
	we stop the programm.

**new_room.c**
	- if the id is bigger than int max, we stop the programm
	execution. It is a security.

+++ src/parsing/get_room.c
@@ -6,7 +6,7 @@
 /*   By: amamy <amamy@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/08/31 16:22:19 by amamy             #+#    #+#             */
-/*   Updated: 2019/10/23 23:10:36 by amamy            ###   ########.fr       */
+/*   Updated: 2019/10/25 22:08:07 by amamy            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -133,7 +133,7 @@ int			get_room(t_room *r, t_farm *f)
 {
 	char	*line;
 	int		ret;
-	int		id;
+	long	id;
 
 	id = 0;
 	ret = gnl_store(0, &line, f, GET_ROOMS);
